move[BE]: Move class files
